feat: add generic  error handling
