+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/BulkActionBar/BulkActionErrorDetails/index.ts
@@ -1,0 +1,1 @@
+export { BulkActionErrorDetails } from './BulkActionErrorDetails.component';
