--- src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/BulkActionBar/BulkActionErrorReports/index.ts
@@ -1,1 +1,0 @@
-export { BulkActionErrorReports } from './BulkActionErrorReports.component';
