+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/BulkActionBar/utils/index.ts
@@ -1,1 +1,2 @@
 export { getRecordUrlFromErrorReport } from './getRecordUrlFromErrorReport';
+export { extractValidationReport } from './extractValidationReport';
